Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/font-shorthand-serialization-font-stretch-expected.txt


FAIL Percentages which can be transformed into keywords should be for serialization assert_equals: expected "ultra-condensed medium serif" but got "50% medium serif"
FAIL Percentages which cannot be transformed into keywords should prevent the font shorthand from serializing assert_equals: expected "" but got "25% medium serif"
